GdMaster init

+++ godot_dir/one_dir/GaalamaBridge/BridgeMain/GdSceneId.cs
@@ -1,3 +1,3 @@
 namespace GaalamaBridge.BridgeMain;
 
-public record GdSceneId(int Val);
+public record GdSceneId(long Val);
